refactor(04): fold object validator onto [SettingsSection].ValidatorType (PR #33 comment #3)
Reuse the existing section attribute for the object-level validator instead of a separate
SettingsValidatorAttribute (now deleted; it was unreleased, so not a breaking change).
ValuesPopulator reads the validator Type from SettingsSectionAttribute.ValidatorType; the
dispatch and value-free exception wrapping are unchanged.

Because [SettingsSection] is the scan-discovery marker, the deliberately-failing validation
fixtures would now be eagerly built by the whole-assembly ScanAssemblies tests. Making them
internal is not viable: the Reflection.Emit proxy generator emits into a random-GUID dynamic
assembly that cannot implement an internal interface (verified: TypeLoadException). So the
fixtures move to a new, never-scanned ExistForAll.SimpleSettings.UnitTests.Fixtures library,
referenced by UnitTests but never passed to ScanAssemblies.

Adds a regression test that the merged validator also runs under the assembly-scan path.
Full suite 143/143 on net10; build clean on net8 + net10.
