+++ src/Core/ExistForAll.SimpleSettings/SettingsSectionAttribute.cs
@@ -5,11 +5,12 @@
 	{
 		public string? Name { get; set; }
 
+		public Type? ValidatorType { get; set; }
+
 		public SettingsSectionAttribute()
 		{
-			
 		}
-		
+
 		public SettingsSectionAttribute(string name)
 		{
 			Name = name;
