+++ src/Tests/ExistForAll.SimpleSettings.UnitTests.Fixtures/ExistForAll.SimpleSettings.UnitTests.Fixtures.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFrameworks>net8.0;net10.0</TargetFrameworks>
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\Core\ExistForAll.SimpleSettings\ExistForAll.SimpleSettings.csproj" />
+  </ItemGroup>
+</Project>
